wip(perf): reduced states in enum; no impact
Search Machine had a 9 variant enum. Tried reudcing it to 7, to see if
fitting within a byte might offer some optimizations.  ... On further
thinking the high level of serialization would likely make that
difficult. (Though the compiler may have noticed some opportuniteis.  It
would be good to have a better sense of register and operation
utilization.)