feat: Ord based on Static OnceLock PageRelations
PartialOrd / Ord only take in the Page values.
We could, reasonably, add an ordering reference to them.  So long as
that reference had Eq/PartialEq defined.  ... and that would probably
be best in most cases.
Instead, we defined a static PageRelations and used that as our
reference for the Ord & PartialOrd `cmp` -- wrapping it in a OnceLock.

(The current code is single threaded, we could use a OnceCell.)