refact: better spantrace usage; test update
A test was inadvertently left on `ignore`.
Taking it off ignore also breaks `cargo test`, due to the shared process
running and use of OnceLock(|Cell).  I'm uncertain if mutex to force
test serialization will solve the problem (as it will with some shared resource
problems) - the OnceLock is set as a static.

Investigation: worth seeing how static drops and test operation work wrt
in the future.

For now, `cargo nextest` is the primary testing tool for non-doctests
already.  I've added a trace `warn!` to alert future users should they
get this error.