wip(turnback): unproductive complexity
Position with Xpos & Ypos could be nice in some situations to prevent
logic errors.  But are creating friction for little value here.  Better
to trust to methods to keep things clean.  (And in theory it wouldn't be
too bad to refactor them in due to the type checking.)

Ergonomic out-of-bounds detection also missing.
Also a proliferation of unneeded traits (particularl from string).