perf: 1_000x difference turning off tracing
~3.3 seconds for Part1_full with it enabled
~3.2 milliseconds with it off.
Running the subscriber only in debug modoe with or without using
a feature flag to scrub tracing code both yield the improvement.
Just scrubing the code, but still running the subscriber
yielded about 4ms (instead of 3.2ms) runtime -- not sure if that's a
fixed cost or if it scales.
I need to set up some different profiling approaches to capture this.